pyproject.toml
[tool.poetry]
name = "ai-for-sales-animals-items"
version = "0.1.0"
description = ""
authors = ["KittGT42 <114111458+KittGT42@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
telethon = "^1.38.1"
openai = "^1.55.0"
python-dotenv = "^1.0.1"
instagrapi = "^2.1.3"
pillow = "^11.0.0"
instabot = "^0.117.0"
flask = {extras = ["async"], version = "^3.1.0"}
pymongo = "^4.10.1"
aioflask = "^0.4.0"
hypercorn = "^0.17.3"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.36"
pydantic-settings = "^2.6.1"
asyncpg = "^0.30.0"
cloudinary = "^1.41.0"
tenacity = "^9.0.0"
aiohttp = "^3.11.11"
django = "^5.1.4"
pandas = "^2.2.3"
anthropic = "^0.43.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
